refactor DTO, fix fr2-2 api endpoint
