+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/controller/DailyTaskController.java
@@ -1,12 +1,11 @@
 package com.mobile.pomodoro.controller;
+
 import com.mobile.pomodoro.dto.request.DailyTaskRequestDTO;
-import com.mobile.pomodoro.dto.request.PlanRequestDTO;
-import com.mobile.pomodoro.dto.request.PlanToEditRequestDTO;
-import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO.DailyTaskResponeseDTO;
+import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO;
 import com.mobile.pomodoro.dto.response.MessageResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO.PlanToEditResponseDTO;
+import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO;
 import com.mobile.pomodoro.entities.User;
-import com.mobile.pomodoro.services.IDailyTaskService;
+import com.mobile.pomodoro.services.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -30,7 +29,7 @@ public class DailyTaskController {
             @RequestAttribute(name = "user") User user) {
         return dailyTaskService.createDailyTask(request, user.getUserId());
     }
-    @GetMapping("/api/v1/daily-task/{id}")
+    @GetMapping("/{id}")
     public ResponseEntity<?> getDailyTaskPlanDetails(
             @PathVariable Long id,
             @RequestAttribute(name = "user") User user) {
@@ -46,16 +45,16 @@ public class DailyTaskController {
             );
         }
     }
-    @PostMapping("/plan-to-edit")
-    public ResponseEntity<PlanToEditResponseDTO> planToEdit(
-            @RequestBody PlanRequestDTO request,
-            @RequestAttribute(name = "user") User user) {
-        try {
-            return new ResponseEntity<>(dailyTaskService.planToEdit(request, user), HttpStatus.OK);
-        } catch (Exception e) {
-            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
-        }
-    }
 
+//    @PostMapping("/plan-to-edit")
+//    public ResponseEntity<?> planToEdit(
+//            @RequestBody PlanRequestDTO request,
+//            @RequestAttribute(name = "user") User user) {
+//        try {
+//            return new ResponseEntity<>(dailyTaskService.planToEdit(request, user), HttpStatus.OK);
+//        } catch (Exception e) {
+//            return new ResponseEntity<>(null, HttpStatus.INTERNAL_SERVER_ERROR);
+//        }
+//    }
 
 }
