+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/controller/PlanController.java
@@ -1,9 +1,8 @@
 package com.mobile.pomodoro.controller;
 
 import com.mobile.pomodoro.dto.request.PlanRequestDTO;
-import com.mobile.pomodoro.dto.response.MessageResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanTaskResponeseDTO.PlanTaskResponeseDTO;
-import com.mobile.pomodoro.dto.response.PlanResponseDTO.PlanResponseDTO;
+import com.mobile.pomodoro.dto.response.PlanTaskResponeseDTO;
+import com.mobile.pomodoro.dto.response.PlanResponseDTO;
 import com.mobile.pomodoro.entities.User;
 import com.mobile.pomodoro.services.IPlanService;
 import lombok.RequiredArgsConstructor;
