+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/dto/response/DailyTaskResponeseDTO.java
@@ -1,7 +1,7 @@
-package com.mobile.pomodoro.dto.response.ToDoResponeseDTO;
+package com.mobile.pomodoro.dto.response;
+
 import lombok.*;
 import lombok.experimental.FieldDefaults;
-
 import java.util.List;
 
 @Data
@@ -9,15 +9,16 @@ import java.util.List;
 @NoArgsConstructor
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class ToDoResponseDTO {
-    List<SingleToDoDTO> list;
+public class DailyTaskResponeseDTO {
+    List<SingleDailyTaskDTO> list;
 
     @Data
     @AllArgsConstructor
     @NoArgsConstructor
     @Builder
     @FieldDefaults(level = AccessLevel.PRIVATE)
-    public static class SingleToDoDTO {
+    public static class SingleDailyTaskDTO {
+        long plan_id;
         String title;
         int is_done;
     }
