+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/dto/response/PlanToEditResponseDTO.java
@@ -1,4 +1,4 @@
-package com.mobile.pomodoro.dto.response.PlanToEditResponseDTO;
+package com.mobile.pomodoro.dto.response;
 import lombok.*;
 import lombok.experimental.FieldDefaults;
 import java.util.List;
