+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/services/IDailyTaskService.java
@@ -2,9 +2,9 @@ package com.mobile.pomodoro.services;
 
 import com.mobile.pomodoro.dto.request.DailyTaskRequestDTO;
 import com.mobile.pomodoro.dto.request.PlanToEditRequestDTO;
-import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO.DailyTaskResponeseDTO;
+import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO;
 import com.mobile.pomodoro.dto.response.MessageResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO.PlanToEditResponseDTO;
+import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO;
 import com.mobile.pomodoro.entities.User;
 
 public interface IDailyTaskService extends IInitializerData{
