+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/services/IPlanService.java
@@ -2,10 +2,8 @@ package com.mobile.pomodoro.services;
 
 import com.mobile.pomodoro.CustomException.UserNotFoundException;
 import com.mobile.pomodoro.dto.request.PlanRequestDTO;
-import com.mobile.pomodoro.dto.response.MessageResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanTaskResponeseDTO.PlanTaskResponeseDTO;
-import com.mobile.pomodoro.dto.response.PlanResponseDTO.PlanResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO.PlanToEditResponseDTO;
+import com.mobile.pomodoro.dto.response.PlanTaskResponeseDTO;
+import com.mobile.pomodoro.dto.response.PlanResponseDTO;
 import com.mobile.pomodoro.entities.User;
 
 public interface IPlanService extends IInitializerData{
