+++ Backend/pomodoro/src/main/java/com/mobile/pomodoro/services/impl/DailyTaskServiceImpl.java
@@ -1,14 +1,12 @@
 package com.mobile.pomodoro.services.impl;
 
-import com.mobile.pomodoro.dto.request.DailyTaskRequestDTO;
-import com.mobile.pomodoro.dto.request.PlanRequestDTO;
+import com.mobile.pomodoro.dto.request.*;
 import com.mobile.pomodoro.dto.request.PlanToEditRequestDTO;
-import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO.DailyTaskResponeseDTO;
+import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO;
+import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO.*;
 import com.mobile.pomodoro.dto.response.MessageResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanResponseDTO.PlanResponseDTO;
-import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO.PlanToEditResponseDTO;
+import com.mobile.pomodoro.dto.response.PlanToEditResponseDTO;
 import com.mobile.pomodoro.entities.DailyTask;
-import com.mobile.pomodoro.dto.response.DailyTaskResponeseDTO.DailyTaskResponeseDTO.SingleDailyTaskDTO;
 import com.mobile.pomodoro.entities.Plan;
 import com.mobile.pomodoro.entities.PlanTask;
 import com.mobile.pomodoro.entities.User;
@@ -16,8 +14,6 @@ import com.mobile.pomodoro.repositories.DailyTaskRepository;
 import com.mobile.pomodoro.repositories.PlanRepository;
 import com.mobile.pomodoro.repositories.PlanTaskRepository;
 import com.mobile.pomodoro.services.IDailyTaskService;
-
-import com.mobile.pomodoro.services.IPlanService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -54,6 +50,7 @@ public class DailyTaskServiceImpl extends AService implements IDailyTaskService 
             if (dailytasks.isEmpty()) {
                 log.warn("Không tìm thấy công việc nào cho userId: {}", userId);
             }
+
             List<SingleDailyTaskDTO> dailytaskDTOs = dailytasks.stream()
                     .map(dailytask -> SingleDailyTaskDTO.builder()
                             .plan_id(dailytask.getPlanId())
@@ -187,6 +184,7 @@ public class DailyTaskServiceImpl extends AService implements IDailyTaskService 
             throw new RuntimeException("Không thể lấy chi tiết DailyTask. " + e.getMessage());
         }
     }
+
     @Override
     public PlanToEditResponseDTO planToEdit(PlanToEditRequestDTO request, User user) {
         log.info(String.format("Bắt đầu xử lý plan-to-edit cho planId: %d và userId: %d", request.getPlan_id(), user.getUserId()));
